\name{cir-package}
\alias{cir-package}
\alias{cir}
\docType{package}
\title{
Isotonic Regression and Centered Isotonic Regression
}
\description{
Centered Isotonic Regression (CIR) is an empirical shrinkage improvement to isotonic regression (IR), requiring no tuning parameters. This package implements CIR for Binomial dose-response data. An IR implementation is trivially included. Interval estimates are also supplied.
}
\details{
\tabular{ll}{
Package: \tab cir\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2014-05-02\cr
License: \tab GPL 3\cr
}
}
\author{
Assaf Oron

Maintainer: Assaf Oron <assaf.oron@seattlechildrens.org>
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
